+++ src/components/connector.vue
@@ -1,7 +1,6 @@
 <script lang="ts">
 import { defineComponent } from 'vue'
-import { wsClient } from '../utils/init';
-import { useI18n } from 'vue-i18n';
+import { wsClient } from '../init/client';
 
 export default defineComponent({
     name: 'connectorComponent',
@@ -25,10 +24,6 @@ export default defineComponent({
             if (!this.websocketURL) return
             wsClient.wsURL = this.websocketURL
         }
-    },
-    setup() {
-        const { t } = useI18n() // use as global scope
-        return { t }
     }
 })
