+++ src/components/status.vue
@@ -1,18 +1,10 @@
 <script lang="ts">
 import {  defineComponent } from 'vue'
-import { useI18n } from 'vue-i18n';
 
 export default defineComponent({
     name: 'statusComponent',
     data: () => ({
-        gcodeCommand: '',
-    }),
-    methods: {
-    },
-    setup() {
-        const { t } = useI18n() // use as global scope
-        return { t }
-    }
+    })
 })
 </script>
