+++ src/init/client.ts
@@ -1,10 +1,8 @@
-import wsConnector from "./wsconnector";
-
+import wsConnector from "../utils/wsconnector";
 
 // WS CLIENT INITIALIZATION
-// get URL from storage if it exists
 const wsURL = localStorage.getItem('wsURL')
-// Create client
+
 export const wsClient = new wsConnector()
-// Define Ws ip
+
 if (wsURL) wsClient.wsURL = wsURL
