+++ src/locale/en.json
@@ -1,0 +1,46 @@
+{
+    "en": {
+        "app": {
+            "connector": "Connector",
+            "files": "Files",
+            "status": "Status"
+        },
+        "files": {
+            "filename": "Filename:",
+            "file_modified_date": "Modified:",
+            "filesize": "Size:"
+        },
+        "status": {
+            "state": "State:",
+            "file": "File:",
+            "elapsed_time": "Elapsed time:",
+            "estimated_time": "Estimated time:"
+        },
+        "connector": {
+            "status": "Connection status:",
+            "connect": "Connect",
+            "disconnect": "Disconnect"
+        },
+        "main": {
+            "control": "Control",
+            "temperature": "Temperature",
+            "terminal": "Terminal",
+            "printerconfig": "Printer Config",
+            "gcodeviewer": "Gcode viewer"
+        },
+        "control": {
+            "movement_value": "Movement value (X,Y,Z)",
+            "extruder_value": "Extruder value (E)",
+            "btn_extrude": "Extrude",
+            "btn_retract": "Retract"
+        },
+        "printerconfig": {
+            "btn_save": "Save",
+            "btn_M115": "Get printer information (M115)"
+        },
+        "temperature": {
+            "grad_extruder": "Extruder",
+            "grad_bed": "Bed"
+        }
+    }
+}
